Modify appium environment.
