+++ requirements.txt
@@ -5,3 +5,4 @@ pytest-instafail
 UnittestZero
 Jinja2
 requests
+Appium-Python-Client
